refactor: rename named component / test files to generic index.tsx / index.test.tsx
