--- src/components/messenger/list/conversation-item/index.ts
@@ -1,1 +1,0 @@
-export * from './conversation-item';
