--- src/components/messenger/list/conversation-list-panel/index.ts
@@ -1,1 +1,0 @@
-export * from './conversation-list-panel';
