--- src/components/messenger/list/create-conversation-panel/index.ts
@@ -1,1 +1,0 @@
-export * from './create-conversation-panel';
