--- src/components/messenger/list/group-details-panel/index.ts
@@ -1,1 +1,0 @@
-export * from './group-details-panel';
