+++ src/components/messenger/list/group-details-panel/index.tsx
