--- src/components/messenger/list/panel-header/index.ts
@@ -1,1 +1,0 @@
-export * from './panel-header';
