--- src/components/messenger/list/selected-user-tag/index.ts
@@ -1,1 +1,0 @@
-export * from './selected-user-tag';
