+++ src/components/messenger/list/start-group-panel/index.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 
-import { StartGroupPanel, Properties } from './start-group-panel';
+import { StartGroupPanel, Properties } from '.';
 
 jest.mock('../../autocomplete-members');
 jest.mock('@zero-tech/zui/components');
