--- src/components/messenger/list/start-group-panel/index.ts
@@ -1,1 +1,0 @@
-export * from './start-group-panel';
