--- src/components/messenger/list/user-search-results/index.ts
@@ -1,1 +1,0 @@
-export * from './user-search-results';
