--- src/components/messenger/list/utils/index.ts
@@ -1,1 +1,0 @@
-export * from './utils';
